🪲 [Fix]: Fix nightly run issues (#66)
## Description

This pull request includes several changes to the `scripts/main.ps1`
file to enhance error handling and improve the script's behavior when it
is not run for a pull request.

Error handling improvements:

* Added a `try` block at the beginning of the script to encapsulate the
main logic and a `catch` block at the end to throw any caught
exceptions.
[[1]](diffhunk://#diff-dc2e5a659836b1b73abb03421c567f5018c2755677c4a0aa764cb26117b68011R11)
[[2]](diffhunk://#diff-dc2e5a659836b1b73abb03421c567f5018c2755677c4a0aa764cb26117b68011R325-R327)

Behavior improvements for non-pull request events:

* Added a check to determine if the event is not a pull request and exit
the script with a warning message if it is not.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [x] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas